refactor: refactor nft collection list
