+++ src/pages/NftCollections/styles.module.scss
@@ -108,10 +108,11 @@
     text-align: center;
   }
 
-  .buttonIcon {
+  .typeFilter {
     cursor: pointer;
     display: flex;
     align-items: center;
+    justify-content: end;
   }
 
   .activeIcon {
@@ -127,55 +128,44 @@
   }
 }
 
-.filterPop {
-  :global {
-    .ant-popover-content {
-      .ant-popover-arrow {
+.filterItems {
+  display: flex;
+  flex-direction: column;
+  width: 200px;
+  a {
+    display: flex;
+    justify-content: space-between;
+    color: var(--primary-color);
+    padding: 10px 0px 10px 10px;
+    margin-right: 8px;
+    border-radius: 8px;
+    &:hover {
+      background: var(--primary-hover-bg-color);
+      color: var(--primary-color);
+      cursor: pointer;
+    }
+    &[data-is-active='false'] {
+      color: #000;
+      svg {
         display: none;
       }
     }
-    .ant-popover-inner-content {
+  }
+  svg path {
+    fill: var(--primary-color);
+  }
+}
+
+.antPopover {
+  :global {
+    .ant-popover-inner {
       border-radius: 8px;
-      padding: 14px 15px 14px 16px;
-
-      @media (max-width: 750px) {
-        padding: 14px 15px 14px 8px !important;
-      }
-
-      > div {
-        width: 200px;
-        > button {
-          margin: 0;
-          border: none;
-          outline: none;
-          width: 100%;
-          background-color: transparent;
-          padding: 10px 0 10px 8px;
-          border-radius: 8px;
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-          align-items: center;
-          &:hover {
-            background: var(--primary-hover-bg-color);
-            color: var(--primary-color);
-            cursor: pointer;
-          }
-          > div:last-child {
-            height: 16px;
-            width: 16px;
-            margin-left: 10px;
-            path {
-              fill: var(--primary-color);
-            }
-          }
-        }
-      }
+      box-shadow: 0px 2px 10px 0px #eeeeee;
     }
   }
 }
 
-.transactions_header {
+.transactionsHeader {
   > span {
     cursor: pointer;
     &.sortActive {
@@ -207,12 +197,54 @@
   }
 }
 
-.holder_and_minted_header {
-  > span {
-    cursor: pointer;
-    &.sortActive {
-      color: var(--primary-color);
+.holderMinted {
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  cursor: pointer;
+  .sortActive {
+    color: var(--primary-color);
+  }
+}
+
+div[data-role='mobile-list'] {
+  display: none;
+  font-size: 0.875rem;
+  .listHeader {
+    display: flex;
+    height: 3rem;
+    & > div {
+      margin-right: 3rem;
+      white-space: nowrap;
     }
+  }
+  dl {
+    margin-left: 36px;
+    margin-bottom: 10px;
+    display: flex;
+    dt {
+      flex: 1;
+      min-width: 105px;
+    }
+    dd {
+      flex: 1;
+      margin: 0;
+    }
+  }
+  .listItem {
+    border-bottom: 1px solid #f0f0f0;
+    margin-bottom: 1.15rem;
+    dt::after {
+      content: ':';
+    }
+    dd {
+      a {
+        font-weight: 500 !important;
+      }
+    }
+  }
+  .loading {
+    text-align: center;
   }
 }
 
@@ -222,18 +254,26 @@
   }
 }
 
-@media screen and (max-width: 790px) {
+@media screen and (max-width: 1000px) {
   .container {
     margin: 40px 8px 60px;
-    th:first-child,
-    th:last-child,
-    td:nth-last-child(2),
-    td:last-child {
+    table[data-role='desktop-list'] {
       display: none;
     }
-    th:nth-child(2),
-    th:nth-child(3) {
-      text-align: left;
+    div[data-role='mobile-list'] {
+      display: block;
     }
+  }
+  .antPopover {
+    :global {
+      .ant-popover-inner-content {
+        padding: 14px 15px 14px 12px !important;
+      }
+    }
+  }
+}
+@media screen and (max-width: 350px) {
+  div[data-role='mobile-list'] {
+    font-size: 0.8rem;
   }
 }
